refactor(core): CATALYST-00 split VariantSelector
