+++ apps/core/components/VariantSelector/CheckboxField.tsx
@@ -1,0 +1,29 @@
+import { CheckboxOption } from '@bigcommerce/catalyst-client';
+import { Checkbox } from '@bigcommerce/reactant/Checkbox';
+import { Label } from '@bigcommerce/reactant/Label';
+
+export const CheckboxField = ({ option }: { option: CheckboxOption }) => (
+  <fieldset>
+    <legend className="inline-block py-2 font-semibold">
+      {option.isRequired ? (
+        <>
+          {option.displayName} <span className="font-normal text-gray-500">(required)</span>
+        </>
+      ) : (
+        option.displayName
+      )}
+    </legend>
+    <div className="flex flex-row items-center" key={option.entityId}>
+      <Checkbox
+        defaultChecked={option.checkedByDefault}
+        id={`${option.entityId}`}
+        name={`attribute[${option.entityId}]`}
+        required={option.isRequired}
+        value={option.checkedOptionValueEntityId}
+      />
+      <Label className="mx-3 font-normal" htmlFor={`${option.entityId}`}>
+        {option.label}
+      </Label>
+    </div>
+  </fieldset>
+);
